Optimizations to speed up code 2x
I removed the print statement, as it isn't really necessary. I moved i1 out of the innermost loop, and replaced np.absolute with abs - this speeds the code up by 2x, and the image is exactly the same. 
I also replaced the enumerate calls with range() calls, as you don't use the rows and columns.