Optimizations to speed up code 2x (#1)
* Optimizations to speed up code 2x

I removed the print statement, as it isn't really necessary. I moved i1 out of the innermost loop, and replaced np.absolute with abs - this speeds the code up by 2x, and the image is exactly the same. 
I also replaced the enumerate calls with range() calls, as you don't use the rows and columns.

* Move abs() out of inner loop for speed up.

Generally, it's best to do as little as possible inside the innermost loop, so I used numpy's abs function outside of the loop for the same effect.

* Change not > to <=

* Add support for generating PNGs natively.

I've added command line args to allow setting the image file name, dimensions, threshold, and random noise seed.  Additionally, removed dependency on PyGame and replaced it with a dependency on PIL to save and display images.

* Remove extraneous imports, allow threshold and seed to be floats.

* Update README.md

Change requirements.

* Update README.md

* Add large and small sample image with better cropping.

* Update demo images

* Update README.md